api/requirements.txt
fastapi
vanna[fastapi,postgres,openai]
psycopg2-binary
python-dotenv
PyJWT
uvicorn
